check files before compiling

+++ config/zsh/dotfiles.zsh
@@ -4,10 +4,10 @@ plugins_home=$ZSH_HOME/plugins
 
 function _dotfiles_benchmark() {
   # download zsh-bench
-  if [[ ! -d $plugins_home/zsh-bench ]]; then
-    git clone --depth 1 https://github.com/romkatv/zsh-bench $plugins_home/zsh-bench
-  else
+  if [[ -d $plugins_home/zsh-bench ]]; then
     git -C $plugins_home/zsh-bench pull
+  else
+    git clone --depth 1 https://github.com/romkatv/zsh-bench $plugins_home/zsh-bench
   fi
 
   # run zsh-bench
@@ -18,31 +18,31 @@ function _dotfiles_reload_shell() {
   exec zsh -l
 }
 
-function _dotfiles_update() {
-  # delete compiled files
-  for file in $(find $DOTFILES_HOME/config/zsh -type f -regex ".*\.zwc"); do
-    rm -f $file
+function _dotfiles_compile() {
+  for file in $(find $DOTFILES_HOME/config/zsh -type f -name "*.zsh"); do
+    [[ $file.zwc -ot $file ]] && rm -f $file.zwc
+    [[ -e $file.zwc ]] || zcompile -R -- $file.zwc $file
   done
+}
 
+function _dotfiles_update() {
   for plugin in $dotfiles_plugins; do
-    plugin_dir=$plugins_home/$(echo ${plugin%/*})_$(basename $plugin)
+    plugin_dir=$plugins_home/$plugin
 
     # remove symlink to init.zsh if it exists
-    if [[ -L $plugin_dir/init.zsh ]]; then
-      rm $plugin_dir/init.zsh
-    fi
+    [[ -L $plugin_dir/init.zsh ]] && rm $plugin_dir/init.zsh
 
     # pull/download plugin repo
-    if [[ ! -d $plugin_dir ]]; then
-      git clone --depth 1 https://github.com/$plugin.git $plugin_dir
-    else
+    if [[ -d $plugin_dir ]]; then
       git -C $plugin_dir pull
+    else
+      git clone --depth 1 https://github.com/$plugin.git $plugin_dir
     fi
 
     # find zsh file for plugin and make symlink if it isn't init.zsh
     if [[ ! -e $plugin_dir/init.zsh ]]; then
       init_candidates=($plugin_dir/*.plugin.{z,}sh(N) $plugin_dir/*.{z,}sh{-theme,}(N))
-      (( $#init_candidates )) || { echo >&2 "No init file found for $plugin." && continue }
+      (( $#init_candidates )) || { echo >&2 "no init file found for $plugin!" && continue }
       ln -sf ${init_candidates[1]} $plugin_dir/init.zsh
     fi
   done
@@ -51,20 +51,16 @@ function _dotfiles_update() {
   git -C $DOTFILES_HOME pull
 
   # rebuild plugins/init.zsh
-  if [[ ! -e $plugins_home/init.zsh ]]; then
-    touch $plugins_home/init.zsh
+  if [[ -e $plugin_dir/init.zsh ]]; then
+    truncate -s 0 "$plugins_home/init.zsh"
   else
-    truncate -s 0 $plugins_home/init.zsh
+    touch "$plugins_home/init.zsh"
   fi
   for plugin in $dotfiles_plugins; do
-    echo "source $plugins_home/$(echo ${plugin%/*})_$(basename $plugin)/init.zsh" >> $plugins_home/init.zsh
+    echo "source $plugins_home/$plugin/init.zsh" >> $plugins_home/init.zsh
   done
 
-  # recompile files
-  for file in $(find $DOTFILES_HOME/config/zsh -type f -regex ".*\.zsh"); do
-    zcompile -R -- $file.zwc $file
-  done
-
+  _dotfiles_compile
   _dotfiles_reload_shell
 }
 
@@ -72,6 +68,7 @@ function _dotfiles_help() {
   echo "Valid usage: dotfiles [ reload | update | benchmark ]"
   echo
   echo "reload: reloads shell with \"exec zsh -l\""
+  echo "compile: recompiles all *.zsh files under config/zsh"
   echo "update: runs \"git pull\" for each plugin repo and the dotfiles repo, recompiles, reloads shell"
   echo "benchmark: runs benchmark using zsh-bench"
 }
@@ -81,6 +78,7 @@ function dotfiles() {
     case "$1" in
       benchmark) _dotfiles_benchmark;;
       reload) _dotfiles_reload_shell;;
+      compile) _dotfiles_compile;;
       update) _dotfiles_update;;
       help) _dotfiles_help;;
       *) echo "bad argument: $1" && echo && _dotfiles_help;;
@@ -90,7 +88,5 @@ function dotfiles() {
 }
 
 # load plugins
-if [[ ! -e $plugins_home/init.zsh ]]; then
-  _dotfiles_update
-fi
+[[ -e $plugins_home/init.zsh ]] || _dotfiles_update
 source $plugins_home/init.zsh
